requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.10
django-phone-verify==0.1.1
django-phonenumber-field==2.4.0
djangorestframework==3.9.4
idna==2.8
phonenumbers==8.10.12
pycparser==2.19
Pygments==2.4.1
PyJWT==1.7.1
PySocks==1.7.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
django-environ==0.4.5
black==19.10b0
isort==4.3.21
